+++ splitio/api/auth.py
@@ -3,7 +3,8 @@
 import logging
 import json
 
-from splitio.api import APIException, headers_from_metadata
+from splitio.api import APIException
+from splitio.api.commons import headers_from_metadata
 from splitio.api.client import HttpClientException
 from splitio.models.token import from_raw
