+++ splitio/api/events.py
@@ -1,8 +1,9 @@
 """Events API module."""
 import logging
 
-from splitio.api import APIException, headers_from_metadata
+from splitio.api import APIException
 from splitio.api.client import HttpClientException
+from splitio.api.commons import headers_from_metadata
 
 
 _LOGGER = logging.getLogger(__name__)
