+++ splitio/api/segments.py
@@ -3,7 +3,8 @@
 import json
 import logging
 
-from splitio.api import APIException, headers_from_metadata, build_fetch
+from splitio.api import APIException
+from splitio.api.commons import headers_from_metadata, build_fetch
 from splitio.api.client import HttpClientException
 
 
@@ -40,7 +41,7 @@ class SegmentsAPI(object):  # pylint: disable=too-few-public-methods
         :type change_number: int
 
         :param fetch_options: Fetch options for getting segment definitions.
-        :type fetch_options: splitio.api.FetchOptions
+        :type fetch_options: splitio.api.commons.FetchOptions
 
         :return: Json representation of a segmentChange response.
         :rtype: dict
