+++ splitio/sync/segment.py
@@ -1,7 +1,8 @@
 import logging
 import time
 
-from splitio.api import APIException, FetchOptions
+from splitio.api import APIException
+from splitio.api.commons import FetchOptions
 from splitio.tasks.util import workerpool
 from splitio.models import segments
 from splitio.util.backoff import Backoff
