+++ splitio/sync/split.py
@@ -5,7 +5,8 @@ import itertools
 import yaml
 import time
 
-from splitio.api import APIException, FetchOptions
+from splitio.api import APIException
+from splitio.api.commons import FetchOptions
 from splitio.models import splits
 from splitio.util.backoff import Backoff
