+++ tests/api/test_segments_api.py
@@ -1,7 +1,9 @@
 """Segment API tests module."""
 
 import pytest
-from splitio.api import segments, client, APIException, FetchOptions
+
+from splitio.api import segments, client, APIException
+from splitio.api.commons import FetchOptions
 from splitio.client.util import SdkMetadata
