+++ tests/sync/test_splits_synchronizer.py
@@ -3,7 +3,8 @@
 import pytest
 
 from splitio.util.backoff import Backoff
-from splitio.api import APIException, FetchOptions
+from splitio.api import APIException
+from splitio.api.commons import FetchOptions
 from splitio.storage import SplitStorage
 from splitio.models.splits import Split
